requirements-app.txt
fastapi>=0.100
uvicorn[standard]>=0.23
PyJWT[crypto]>=2.8
cryptography>=41
